Redefine elog entry.

+++ arch/arm/mach-msm/htc_battery_vivo.c
@@ -1846,11 +1846,11 @@ static int __init htc_battery_init(void)
 }
 
 // elogk
-void get_fresh_batt_info(unsigned int *fresh_vol, unsigned int *fresh_curr)
+void get_fresh_batt_info(unsigned short *fresh_vol, unsigned short *fresh_curr)
 {
     update_batt_info();
-    *fresh_vol = htc_batt_info.rep.batt_vol;
-    *fresh_curr = htc_batt_info.rep.batt_current;
+    *fresh_vol = (unsigned short)htc_batt_info.rep.batt_vol;
+    *fresh_curr = (unsigned short)htc_batt_info.rep.batt_current;
     return;
 }
 // elogk
